+++ src/components/charts/scatter/ScatterChartPro/ScatterChartPro.emb.ts
@@ -1,0 +1,8 @@
+import { defineComponent } from '@embeddable.com/react';
+import { scatterChartPro } from './definition';
+
+export const preview = scatterChartPro.preview;
+
+export const meta = scatterChartPro.meta;
+
+export default defineComponent(scatterChartPro.Component, meta, scatterChartPro.config);